wavget track mode
track mode to display silence/voice sequences (for gain configuration).
no file operation: 'python -d stt_engines/snowboy/mainwav.py <gain> track'

+++ stt_engines/snowboy/wavget.py
@@ -10,16 +10,15 @@ import logging
 import glob
 
 logging.basicConfig()
-logger = logging.getLogger("snowboy")
+logger = logging.getLogger("jarvis")
 logger.setLevel(logging.INFO)
 TOP_DIR = os.path.dirname(os.path.abspath(__file__))
 RESOURCE_FILE = os.path.join(TOP_DIR, "resources","common.res")
 
-class WaveGet(object):
+class WavGet(object):
     """
     Snowboy decoder to save a wave.
 
-    :param sensitivity: decoder sensitivity, a float.
     :param audio_gain: multiply input volume by this factor.
     :param trigger_ticks: ticks before triggering callback, tick is a sleep_time.
                    [0] ticks_silence_before_detect:
@@ -38,7 +37,7 @@ class WaveGet(object):
             play_data = chr(0) * len(in_data)
             return play_data, pyaudio.paContinue
 
-        a_model=glob.glob( os.path.join(TOP_DIR,"resources","*.pmdl") );
+        a_model=glob.glob( os.path.join(TOP_DIR,"resources","*.[up]mdl") );
         assert len(a_model) > 0, "Need at least one model in resources to proceed"
 
         self.detector = snowboydetect.SnowboyDetect(
@@ -46,7 +45,7 @@ class WaveGet(object):
             model_str=a_model[0].encode())
         self.detector.SetAudioGain( int(audio_gain) )
         """ match or not - it does not matter """
-        self.detector.SetSensitivity("0.1".encode())
+        self.detector.SetSensitivity("0.01".encode())
 
         self.trigger_ticks = trigger_ticks
 
@@ -64,16 +63,14 @@ class WaveGet(object):
 
     def start(self,
               output_file,
+              track_mode,
               interrupt_check=lambda: False,
               sleep_time=0.03):
         """
         Start the voice detector. For every `sleep_time` second it checks the
-        audio buffer for triggering keywords. If detected, then call
-        corresponding function in `detected_callback`, which can be a single
-        function (single model) or a list of callback functions (multiple
-        models). Every loop it also calls `interrupt_check` -- if it returns
-        True, then breaks from the loop and return.
+        audio buffer.
 
+        :param output_file: output file 'wav'. Unlink if failed.
         :param interrupt_check: a function that returns True if the main loop
                                 needs to stop.
         :param float sleep_time: how much time in second every loop waits.
@@ -83,14 +80,24 @@ class WaveGet(object):
             logger.debug("detect voice return")
             return
 
-        tticks = self.trigger_ticks
-        fh = open( output_file + ".raw", 'wb' )
+        if track_mode is None:
+            track_mode = False
+
+        tticks = None
+        fh = None
+        if track_mode == False:
+            if os.path.isfile(output_file):
+                os.unlink(output_file)
+            tticks = self.trigger_ticks
+            fh = open( output_file + ".raw", 'wb' )
+
         """
         to avoid 'truncate' system call.
         also the return status
         """
         fh_writen = False
 
+        cnt = 0
         silence_before = 0
         voice = 0
         silence_after = 0
@@ -107,7 +114,20 @@ class WaveGet(object):
                 continue
 
             ans = self.detector.RunDetection(data)
+            cnt += 1
 
+            """ track mode """
+            if track_mode:
+                fmt = "%6d " % (cnt)
+                if ans == -1:
+                    logger.warning(fmt+"Error initializing streams or reading audio data")
+                elif ans == -2:
+                    logger.warning(fmt+"Silence")
+                elif ans >= 0:
+                    logger.warning(fmt+"Voice")
+                continue
+
+            """ store file mode """
             if ans == -1:
                 logger.warning("Error initializing streams or reading audio data")
 
@@ -150,25 +170,26 @@ class WaveGet(object):
 
             data_onetick_before = data
 
-        fh.close()
+        if track_mode == False:
+            fh.close()
 
-        logger.warning("Ticks status: " +
-                `silence_before` + " " +
-                `voice` + " " +
-                `silence_after` + " - return: " + `fh_writen`)
+            logger.warning("Ticks status: " +
+                    `silence_before` + " " +
+                    `voice` + " " +
+                    `silence_after` + " - return: " + `fh_writen`)
 
-        logger.warning("Error initializing streams or reading audio data")
-        logger.debug("finished.")
+            logger.warning("Error initializing streams or reading audio data")
+            logger.debug("finished.")
 
-        if fh_writen == True:
-            """ convert to wave """
-            from subprocess import call
-            call([
-                "sox","-r","16000","-c","1","-b","16","-e","signed-integer",
-                "-t","raw",output_file + ".raw",
-                "-t","wav",output_file])
+            if fh_writen == True:
+                """ convert to wave """
+                from subprocess import call
+                call([
+                    "sox","-r","16000","-c","1","-b","16","-e","signed-integer",
+                    "-t","raw",output_file + ".raw",
+                    "-t","wav",output_file])
 
-        os.unlink(output_file + ".raw")
+            os.unlink(output_file + ".raw")
 
         return fh_writen
